Moved AggregatedComponentFrame.py and TriggerWindow into guiutils.py
