+++ guiutils.py
@@ -9,8 +9,7 @@
 # WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
 # PARTICULAR PURPOSE. See the GNU General Public License for more details.
 
-This file contains helper functions for the ESMB gui, so that they don't need to be rewritten,
-    or worse, imported from the gui components themselves
+This file contains helper functions and custom widgets for the ESMB gui
 
 '''
 from tkinter import *
@@ -34,6 +33,180 @@ def buildMandOptFrame(parent, subComponentName, numMandatory, numOptionals, list
 
     return newFrame
 #end buildMandOptFrame
+
+
+class AggregatedComponentFrame(ttk.Frame):
+
+    def __init__(self, app, parent, sectionName, componentType):
+        ttk.Frame.__init__(self, parent)
+
+        self.app           = app
+        self.sectionName   = sectionName
+        self.componentType = componentType
+        self.componentList = []
+
+        self.outer = ttk.Frame(self)
+        self.outer.pack()
+
+        sectionNameLabel = ttk.Label(self.outer, text=self.sectionName, anchor="center")
+        sectionNameLabel.pack()
+
+        self.inner = ttk.Frame(self.outer)
+        self.inner.pack()
+
+        buttonText = "Add " + self.componentType
+        addButton = ttk.Button(self.outer, text=buttonText, width=31, command=self.__addComponent)
+        addButton.pack(expand=True, fill="x")
+    #end init
+
+
+    def __addComponent(self):
+        print("Adding %s to %s..." % (self.componentType, self.sectionName))
+
+        ComponentFrame(self)
+
+        if self.componentType is "trigger":
+            self.componentList[-1].missionComponent = self.app.activeMission.addTrigger()
+            #print(self.componentList[-1].missionComponent)
+            self.editComponent(self.componentList[-1])
+
+        print("Done.")
+    #end __addComponent
+
+
+    def deleteComponent(self, component):
+        print("Removing %s from %s..." % (self.componentType, self.sectionName))
+
+        if self.componentType is "trigger":
+            self.app.activeMission.removeTrigger(component.missionComponent)
+        self.componentList.remove(component)
+        component.pack_forget()
+        component.destroy()
+
+        print("Done.")
+    #end deleteComponent
+
+
+    def editComponent(self, component):
+        print("Editing ", end="")
+        print(component.missionComponent)
+        TriggerWindow(self.app, self.app.gui, component.missionComponent)
+    #end editComponent
+
+#end class AggregatedComponentFrame
+
+
+class ComponentFrame(object):
+
+    def __init__(self, master):
+        self.missionComponent = None
+
+        componentFrame = ttk.Frame(master.inner)
+        componentFrame.pack()
+
+        #labelText = "%s %d" % (master.componentType, len(master.componentList))
+        label = ttk.Label(componentFrame, text=master.componentType, anchor="w")
+        label.grid(row=0, column=0, sticky="ew")
+
+        master.componentList.append(componentFrame)
+
+        editButton = ttk.Button(componentFrame, text="edit", command=lambda: master.editComponent(componentFrame))
+        editButton.grid(row=0, column=1)
+
+        deleteButton = ttk.Button(componentFrame, text="X", command=lambda: master.deleteComponent(componentFrame))
+        deleteButton.grid(row=0, column=2)
+    #end init
+
+#end class ComponentFrame
+
+
+class TriggerWindow(object):
+
+    def __init__(self, app, master, trigger):
+        print("\tBuilding TriggerWindow...", end="\t\t")
+
+        self.app = app
+        self.trigger = trigger
+
+        self.top = Toplevel(master)
+        self.top.title("Edit Trigger")
+        self.top.configure(bg="#ededed")
+        self.top.grab_set()  # freezes the app until the user enters or cancels
+
+        outer = ttk.Frame(self.top)
+        outer.pack(side=TOP)
+
+        self.leftFrame = ttk.Frame(outer)
+        self.leftFrame.pack(side=LEFT)
+
+        self.rightFrame = ttk.Frame(outer)
+        self.rightFrame.pack(side=RIGHT)
+
+        self.closeButton = ttk.Button(self.top, text="Ok", command=self.cleanup)
+        self.closeButton.pack(side=BOTTOM)
+
+        # declare all the variables in one place
+
+        #TODO: find a way to support "on enter <system>"
+        self.action = None
+        actionsList = ["offer", "complete", "accept", "decline", "defer", "fail", "visit", "stopover"]
+
+
+
+        ### BUILDING LEFT FRAME ###
+
+        ## on action
+        onLabel = ttk.Label(self.leftFrame, text="on", width=6)
+        onLabel.grid(row=0, column=0, sticky="w", padx=(5,0))
+
+        self.onActionCombobox = ttk.Combobox(self.leftFrame, state="readonly", values=actionsList)
+        self.onActionCombobox.bind("<<ComboboxSelected>>", self.actionSelected)
+        self.onActionCombobox.grid(row=0, column=1, sticky="ew")
+
+        self.dialogSubComponent = buildMandOptFrame(self.leftFrame, "dialog", 1, 0, ["<text>"])
+        self.dialogSubComponent.grid(row=1, column=0, columnspan=2, sticky="ew")
+
+        self.outfitSubComponent = buildMandOptFrame(self.leftFrame, "outfit", 1, 1, ["<outfit>", "[<number#>]"])
+        self.outfitSubComponent.grid(row=2, column=0, columnspan=2, sticky="ew")
+
+        self.requireSubComponent = buildMandOptFrame(self.leftFrame, "require", 1, 1, ["<outfit>", "[<number#>]"])
+        self.requireSubComponent.grid(row=3, column=0, columnspan=2, sticky="ew")
+
+        self.paymentSubComponent = buildMandOptFrame(self.leftFrame, "payment", 0, 2, ["<base#>", "[<multiplier#>]"])
+        self.paymentSubComponent.grid(row=4, column=0, columnspan=2, sticky="ew")
+
+        self.eventSubComponent = buildMandOptFrame(self.leftFrame, "event", 1, 2, ["<name>", "[<delay#>]", "[<max#>]"])
+        self.eventSubComponent.grid(row=5, column=0, columnspan=2, sticky="ew")
+
+        self.failSubComponent = buildMandOptFrame(self.leftFrame, "fail", 0, 1, ["[<name>]"])
+        self.failSubComponent.grid(row=6, column=0, columnspan=2, sticky="ew")
+
+        self.logs = AggregatedComponentFrame(self.app, self.leftFrame, "Logs", "log")
+        self.logs.grid(row=7, column=0, columnspan=2, sticky="ew")
+
+        ### DONE BUILDING LEFT FRAME ###
+
+
+        # build the right frame
+        testR = ttk.Label(self.rightFrame, text="RightSideFrame")
+        testR.pack()
+
+        print("Done.")
+    #end init
+
+
+    def actionSelected(self, event):
+        self.action = self.onActionCombobox.get()
+        print('\nTrigger action selected: "on %s"' % self.action)
+    #end actionSelected
+
+
+    def cleanup(self):
+        self.top.grab_release()  # HAVE TO RELEASE
+        self.top.destroy()
+    #end cleanup
+
+#end class TriggerWindow
 
 
 class _SubComponentMandOptFrame(ttk.Frame):
